made all files contain only 5 functions, updated Makefile, updated README and header to incorporate the file changes
