+++ builtinA.c
@@ -111,42 +111,3 @@ int _monalisa(arg_inventory_t *arginv)
 
 	return (EXT_SUCCESS);
 }
-
-/**
- * the_help - prints mona lisa ascii art
- * @arginv: arguments inventory
- *
- * Return: 0 on success
- */
-int the_help(arg_inventory_t *arginv)
-{
-	char **commands;
-	int i = 0;
-	bins_t bins[] = {
-		{"exit", h_exit}, {"monalisa", h_monalisa}, {"env", h_env},
-		{"setenv", h_setenv}, {"unsetenv", h_unsetenv},
-		{"history", h_history}, {"cd", h_cd}, {"alias", h_alias},
-		{"help", h_help},
-		{NULL, NULL}
-	};
-
-
-	commands = (char **)arginv->commands;
-	if (commands[2] != NULL)
-	{
-		perror("help: too many input commands.");
-		return (-1);
-	}
-
-	while (bins[i].function != NULL)
-	{
-		if (_strcmp(bins[i].function, commands[1]) == 0)
-		{
-			bins[i].help();
-			break;
-		}
-		i++;
-	}
-
-	return (EXT_SUCCESS);
-}
