+++ free.c
@@ -26,7 +26,7 @@ int freeall(arg_inventory_t *arginv)
 
 /**
  * free_alias - function to free all allocated memory
- * @head - head of alias
+ * @head: head of alias
  *
  * Return: 0 on success, 1 on failure
  */
@@ -48,7 +48,7 @@ int free_alias(alias_t *head)
 
 /**
  * free_environ - function to free all allocated memory
- * @head - head of custom _environ
+ * @head: head of custom _environ
  *
  * Return: 0 on success, 1 on failure
  */
