+++ helpA.c
@@ -1,53 +1,49 @@
 #include "header.h"
 
+/**
+ * h_alias - help function to explain how the associated function works
+ */
 void h_alias(void)
 {
 	_puts("this is a help file:");
 	_puts("first: goolge your question");
 	_puts("second: contact developers for help");
 }
+
+/**
+ * h_cd - help function to explain how the associated function works
+ */
 void h_cd(void)
 {
 	_puts("this is a help file:");
 	_puts("first: goolge your question");
 	_puts("second: contact developers for help");
 }
+
+/**
+ * h_env - help function to explain how the associated function works
+ */
 void h_env(void)
 {
 	_puts("this is a help file:");
 	_puts("first: goolge your question");
 	_puts("second: contact developers for help");
 }
+
+/**
+ * h_exit - help function to explain how the associated function works
+ */
 void h_exit(void)
 {
 	_puts("this is a help file:");
 	_puts("first: goolge your question");
 	_puts("second: contact developers for help");
 }
+
+/**
+ * h_help - help function to explain how the associated function works
+ */
 void h_help(void)
-{
-	_puts("this is a help file:");
-	_puts("first: goolge your question");
-	_puts("second: contact developers for help");
-}
-void h_history(void)
-{
-	_puts("this is a help file:");
-	_puts("first: goolge your question");
-	_puts("second: contact developers for help");
-}
-void h_monalisa(void)
-{	_puts("this is a help file:");
-	_puts("first: goolge your question");
-	_puts("second: contact developers for help");
-}
-void h_setenv(void)
-{
-	_puts("this is a help file:");
-	_puts("first: goolge your question");
-	_puts("second: contact developers for help");
-}
-void h_unsetenv(void)
 {
 	_puts("this is a help file:");
 	_puts("first: goolge your question");
