+++ helpB.c
@@ -1,0 +1,40 @@
+#include "header.h"
+
+/**
+ * h_history - help function to explain how the associated function works
+ */
+void h_history(void)
+{
+	_puts("this is a help file:");
+	_puts("first: goolge your question");
+	_puts("second: contact developers for help");
+}
+
+/**
+ * h_monalisa - help function to explain how the associated function works
+ */
+void h_monalisa(void)
+{	_puts("this is a help file:");
+	_puts("first: goolge your question");
+	_puts("second: contact developers for help");
+}
+
+/**
+ * h_setenv - help function to explain how the associated function works
+ */
+void h_setenv(void)
+{
+	_puts("this is a help file:");
+	_puts("first: goolge your question");
+	_puts("second: contact developers for help");
+}
+
+/**
+ * h_unsetenv - help function to explain how the associated function works
+ */
+void h_unsetenv(void)
+{
+	_puts("this is a help file:");
+	_puts("first: goolge your question");
+	_puts("second: contact developers for help");
+}
