+++ stringsA.c
@@ -1,0 +1,116 @@
+#include "header.h"
+
+/**
+ * _strncat - concatenates from src string to dest string
+ * @dest: destination string
+ * @src: source string
+ * @n: number of bytes to concatenate
+ *
+ * Return: pointer to destination
+ */
+char *_strncat(char *dest, char *src, int n)
+{
+	int i = 0, j = 0;
+
+	while (dest[i] != '\0')
+		i++;
+
+	while (j < n && src[j] != '\0')
+	{
+		dest[i] = src[j];
+		i++;
+		j++;
+	}
+
+	dest[i] = '\0';
+
+	return (dest);
+}
+
+/**
+ * _strcmp - compares string
+ * @s1: first string
+ * @s2: second string
+ *
+ * Return: difference between two ascii valuves
+ */
+
+int _strcmp(const char *s1, const char *s2)
+{
+	int i;
+
+	i = 0;
+
+	while (TRUE)
+	{
+		if (s1[i] != s2[i])
+			return (s1[i] - s2[i]);
+
+		if (s1[i] == '\0' || s2[i] == '\0')
+			break;
+
+		i++;
+	}
+
+	return (0);
+}
+
+/**
+ * _strlen - finds and returns length of string
+ * @str: string to find length
+ *
+ * Return: length of string
+ */
+unsigned int _strlen(const char *str)
+{
+	int i = 0;
+
+	while (str[i] != '\0')
+		i++;
+
+	return (i);
+}
+
+/**
+ * _strdup - takes a string and copies to another a new memory location
+ * @str: string to copy
+ *
+ * Return: pointer to copied string
+ */
+char *_strdup(char *str)
+{
+	unsigned int len, j;
+	char *ptrstring;
+
+	if (str == NULL)
+		return (NULL);
+
+	len = _strlen(str);
+
+	ptrstring = safe_malloc((len + 1) * sizeof(char));
+
+	for (j = 0; j < len; j++)
+		ptrstring[j] = str[j];
+	ptrstring[j] = '\0';
+
+	return (ptrstring);
+}
+
+/**
+ * _strncmp - checks if 2 strings are of equal value and length
+ * @s1: first string
+ * @s2: second string
+ * @n: number of bytes to compare
+ *
+ * Return: difference of first characters that are of diff value or 0 on success
+ */
+int _strncmp(char *s1, char *s2, unsigned int n)
+{
+	unsigned int j;
+
+	for (j = 0; j < n; j++)
+		if (s1[j] != s2[j])
+			return (s1[j] - s2[j]);
+
+	return (0);
+}
